security: don't expose 2FA tokens to WinAuth developer
The tokens get "backed up" to the registry, encrypted with the WinAuth
developer's public PGP key. This is unnecessary behavior, as everyone
using WinAuth expects that these tokens are stored only in the encrypted
database and will not be exposed in any way without the appropriate
user-specified keys.

Additionally, when a diagnostic report is created, all of the 2FA tokens
get appended to the report, and the entire report gets encrypted by the
WinAuth developer's public PGP key.

As far as I'm concerned, the WinAuth developer has no business knowing
my 2FA token details.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>